+++ lib/onepref.dart
@@ -1,3 +1,7 @@
 library onepref;
 
 export 'src/onepref.dart';
+export 'src/iApEngine.dart';
+export 'src/product_id.dart';
+export 'src/benefit.dart';
+export 'src/on_click_animation.dart';